integration/requirements.txt
grpcio==1.46.0
grpcio-tools==1.46.0
loguru==0.6.0